Update polybob tuning.h to latest ddnet version

+++ src/ddnet/polybob/game/tuning.h
@@ -4,16 +4,17 @@
 // This file can be included several times.
 
 #ifndef MACRO_TUNING_PARAM
+// This helps IDEs properly syntax highlight the uses of the macro below.
 #define MACRO_TUNING_PARAM(Name, ScriptName, Value, Description)
 #endif
 
 // physics tuning
 MACRO_TUNING_PARAM(GroundControlSpeed, ground_control_speed, 10.0f, "Max speed the tee can get on ground")
-MACRO_TUNING_PARAM(GroundControlAccel, ground_control_accel, 100.0f / SERVER_TICK_SPEED, "Acceleration speed on the ground")
+MACRO_TUNING_PARAM(GroundControlAccel, ground_control_accel, 100.0f / (float)SERVER_TICK_SPEED, "Acceleration speed on the ground")
 MACRO_TUNING_PARAM(GroundFriction, ground_friction, 0.5f, "Friction on the ground")
 MACRO_TUNING_PARAM(GroundJumpImpulse, ground_jump_impulse, 13.2f, "Impulse when jumping on ground")
 MACRO_TUNING_PARAM(AirJumpImpulse, air_jump_impulse, 12.0f, "Impulse when jumping in air")
-MACRO_TUNING_PARAM(AirControlSpeed, air_control_speed, 250.0f / SERVER_TICK_SPEED, "Max speed the tee can get in the air")
+MACRO_TUNING_PARAM(AirControlSpeed, air_control_speed, 250.0f / (float)SERVER_TICK_SPEED, "Max speed the tee can get in the air")
 MACRO_TUNING_PARAM(AirControlAccel, air_control_accel, 1.5f, "Acceleration speed in air")
 MACRO_TUNING_PARAM(AirFriction, air_friction, 0.95f, "Friction in the air")
 MACRO_TUNING_PARAM(HookLength, hook_length, 380.0f, "Length of the hook")
@@ -49,7 +50,7 @@ MACRO_TUNING_PARAM(LaserDamage, laser_damage, 5, "(UNUSED) Laser damage")
 MACRO_TUNING_PARAM(PlayerCollision, player_collision, 1, "Enable player collisions")
 MACRO_TUNING_PARAM(PlayerHooking, player_hooking, 1, "Enable player vs player hooking")
 
-// ddnet tuning
+//ddnet tuning
 MACRO_TUNING_PARAM(JetpackStrength, jetpack_strength, 400.0f, "Jetpack pistol strength")
 MACRO_TUNING_PARAM(ShotgunStrength, shotgun_strength, 10.0f, "Shotgun pull strength")
 MACRO_TUNING_PARAM(ExplosionStrength, explosion_strength, 6.0f, "Explosion strength (grenade for example)")
